+++ src/uqtestfuns/test_functions/genz.py
@@ -5,15 +5,13 @@ Genz [1] introduced six challenging parameterized $M$-dimensional functions
 designed to test the performance of numerical integration routines:
 
 - Corner peak function features a prominent peak in one corner
-  of the multidimensional space.
+  of the multidimensional space. This function was featured as a test function
+  in sensitivity analysis ([2]) and metamodeling [3] exercises.
 
 The functions are further characterized by shift and scale parameters.
 While the shift parameter has minimal impact on the integral's value,
 the scale parameter significantly affects
 the difficulty of the integration problem.
-
-Some of the functions were featured as test functions in sensitivity
-analysis literature, e.g., [2].
 
 References
 ----------
@@ -26,6 +24,10 @@ References
    global sensitivity analysis,” Reliability Engineering & System Safety,
    vol. 121, pp. 164–174, 2014.
    DOI: 10.1016/j.ress.2013.07.010
+3. J. D. Jakeman, M. S. Eldred, and K. Sargsyan, “Enhancing l1-minimization
+   estimates of polynomial chaos expansions using basis selection,”
+   Journal of Computational Physics, vol. 289, pp. 18–34, 2015.
+   DOI: 10.1016/j.jcp.2015.02.025.
 """
 
 import numpy as np
@@ -37,7 +39,7 @@ from uqtestfuns.core.custom_typing import (
 )
 from uqtestfuns.core.uqtestfun_abc import UQTestFunVarDimABC
 
-__all__ = ["GenzCornerPeak"]
+__all__ = ["GenzCornerPeak", "GenzProductPeak"]
 
 
 MARGINALS_GENZ1984: MarginalSpecs = [
@@ -49,14 +51,19 @@ MARGINALS_GENZ1984: MarginalSpecs = [
     }
 ]
 
-COMMON_TAGS = ["integration", "sensitivity"]
-
 
 def _get_aa_genz_1984(input_dimension: int) -> np.ndarray:
-    """Construct a scaling array for the Genz function from [1]."""
+    """Construct an array of shape parameters for the Genz functions [1]."""
     aa = 5.0 * np.ones(input_dimension)
 
     return aa
+
+
+def _get_bb_genz_1984(input_dimension: int) -> np.ndarray:
+    """Construct an array of shape parameters for the Genz functions [1]."""
+    bb = 0.5 * np.ones(input_dimension)
+
+    return bb
 
 
 def _get_aa_zhang_2014_1(input_dimension: int) -> np.ndarray:
@@ -96,7 +103,7 @@ def evaluate_corner_peak(xx: np.ndarray, aa: np.ndarray) -> np.ndarray:
         M-Dimensional input values given by an N-by-M array where
         N is the number of input values.
     aa : np.ndarray
-        The vector of scale parameters of the Genz family of integrand
+        The vector of shape parameters of the Genz family of integrand
         functions; the larger the value the more difficult the integrations.
 
     Returns
@@ -115,7 +122,7 @@ def evaluate_corner_peak(xx: np.ndarray, aa: np.ndarray) -> np.ndarray:
 class GenzCornerPeak(UQTestFunVarDimABC):
     """A concrete implementation of the corner peak from Genz (1984)."""
 
-    _tags = COMMON_TAGS
+    _tags = ["integration", "metamodeling", "sensitivity"]
     _description = "Corner peak integrand from Genz (1984)"
     _available_inputs: ProbInputSpecs = {
         "Genz1984": {
@@ -131,14 +138,15 @@ class GenzCornerPeak(UQTestFunVarDimABC):
         "Genz1984": {
             "function_id": "GenzCornerPeak",
             "description": (
-                "Parameter set for the corner peak function from Genz (1984)"
+                "Parameter set for the corner peak function from Genz (1984); "
+                "constant shape and offset values"
             ),
             "declared_parameters": [
                 {
                     "keyword": "aa",
                     "value": _get_aa_genz_1984,
                     "type": np.ndarray,
-                    "description": "Scaling parameter",
+                    "description": "Shape parameter",
                 },
             ],
         },
@@ -146,14 +154,14 @@ class GenzCornerPeak(UQTestFunVarDimABC):
             "function_id": "GenzCornerPeak",
             "description": (
                 "Parameter set for the corner peak function from Zhang and "
-                "Pandey (2014); Section 4.4 (3D case)"
+                "Pandey (2014); Section 4.4 (3D case), linearly increasing"
             ),
             "declared_parameters": [
                 {
                     "keyword": "aa",
                     "value": _get_aa_zhang_2014_1,
                     "type": np.ndarray,
-                    "description": "Scaling parameter",
+                    "description": "Shape parameter",
                 },
             ],
         },
@@ -161,14 +169,14 @@ class GenzCornerPeak(UQTestFunVarDimABC):
             "function_id": "GenzCornerPeak",
             "description": (
                 "Parameter set for the corner peak function from Zhang and "
-                "Pandey (2014); Section 4.4 (10D case)"
+                "Pandey (2014); Section 4.4 (10D case), linearly increasing"
             ),
             "declared_parameters": [
                 {
                     "keyword": "aa",
                     "value": _get_aa_zhang_2014_2,
                     "type": np.ndarray,
-                    "description": "Scaling parameter",
+                    "description": "Shape parameter",
                 },
             ],
         },
@@ -176,3 +184,77 @@ class GenzCornerPeak(UQTestFunVarDimABC):
     _default_parameters_id = "Genz1984"
 
     evaluate = staticmethod(evaluate_corner_peak)  # type: ignore
+
+
+def evaluate_product_peak(
+    xx: np.ndarray, aa: np.ndarray, bb: np.ndarray
+) -> np.ndarray:
+    """Compute the product peak function on a set of input values.
+
+    Parameters
+    ----------
+    xx : np.ndarray
+        M-Dimensional input values given by an N-by-M array where
+        N is the number of input values.
+    aa : np.ndarray
+        The vector of shape parameters of the Genz family of integrand
+        functions; the larger the value the more difficult the integrations.
+        The length of the vector must M.
+    bb : np.ndarray
+        The vector of offset parameters of the Genz family of integrand
+        functions; the values do not alter significantly the difficulty of
+        the integration problem.
+        The length of the vector must be M.
+
+    Returns
+    -------
+    np.ndarray
+        The output of the test function evaluated on the input values.
+        The output is a 1-dimensional array of length N.
+    """
+    yy = np.prod(1 / ((xx - bb)**2 + aa**(-2)), axis=1)
+
+    return yy
+
+
+class GenzProductPeak(UQTestFunVarDimABC):
+    """A concrete implementation of the product peak from Genz (1984)."""
+
+    _tags = ["integration"]
+    _description = "Product peak integrand from Genz (1984)"
+    _available_inputs: ProbInputSpecs = {
+        "Genz1984": {
+            "function_id": "GenzProductPeak",
+            "description": (
+                "Input specification for the Genz family of integrand"
+            ),
+            "marginals": MARGINALS_GENZ1984,
+            "copulas": None,
+        }
+    }
+    _available_parameters: FunParamSpecs = {
+        "Genz1984": {
+            "function_id": "GenzProductPeak",
+            "description": (
+                "Parameter set for the product peak function from Genz (1984);"
+                " constant shape and offset values"
+            ),
+            "declared_parameters": [
+                {
+                    "keyword": "aa",
+                    "value": _get_aa_genz_1984,
+                    "type": np.ndarray,
+                    "description": "Shape parameter",
+                },
+                {
+                    "keyword": "bb",
+                    "value": _get_bb_genz_1984,
+                    "type": np.ndarray,
+                    "description": "Offset parameter",
+                },
+            ],
+        },
+    }
+    _default_parameters_id = "Genz1984"
+
+    evaluate = staticmethod(evaluate_product_peak)  # type: ignore